@@ -20,7 +20,6 @@
     "@types/mdx": "^2.0.13",
     "@vercel/analytics": "^1.1.1",
     "clsx": "^1.2.1",
-    "contentlayer": "^0.3.3",
     "date-fns": "^2.29.3",
     "feed": "^4.2.2",
     "framer-motion": "^7.6.19",
@@ -48,6 +47,9 @@
     "unified": "^10.1.2"
   },
   "devDependencies": {
+    "@content-collections/core": "^0.10.0",
+    "@content-collections/markdown": "^0.1.4",
+    "@content-collections/next": "^0.2.6",
     "@eslint/eslintrc": "^3.3.0",
     "@next/eslint-plugin-next": "^15.2.1",
     "@tailwindcss/postcss": "^4.0.14",
@@ -65,7 +67,8 @@
     "postcss": "^8.4.18",
     "prettier": "^3.5.3",
     "tailwindcss": "^4.0.14",
-    "typescript": "4.8.4"
+    "typescript": "5.8.3",
+    "zod": "^4.0.5"
   },
   "peerDependencies": {
     "esbuild": "0.17.x || 0.18.x"
