@@ -1,5 +1,5 @@
-const { withContentlayer } = require('next-contentlayer')
-const withMDX = require('@next/mdx')()
+import { withContentCollections } from '@content-collections/next'
+import createMDX from '@next/mdx'
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
@@ -12,4 +12,6 @@ const nextConfig = {
   },
 }
 
-module.exports = withContentlayer(withMDX(nextConfig))
+const withMDX = createMDX()
+
+export default withContentCollections(withMDX(nextConfig))
