@@ -1,7 +1,6 @@
+import { Post } from 'content-collections'
 import { format } from 'date-fns'
 import Link from 'next/link'
-
-import { Post } from '.contentlayer/generated'
 
 type Props = {
   post: Post
